[#156] refactor: 메인페이지 컴포넌트를 components/main으로 분리
