+++ src/pages/main/Mainpage.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import LetterCarousel from '@/components/letters/LetterCarousel';
 import QuestionCard from '@/components/common/QuestionCard';
-import WriteLetterButtons from './WriteLetterButtons';
-import LetterJourney from './LetterJourney';
+import WriteLetterButtons from '@/components/main/WriteLetterButtons';
+import LetterJourney from '@/components/main/LetterJourney';
 import MainPageSkeleton from '@/components/skeleton/MainPageSkeleton';
 import { useHomeSummary } from '@/hooks/useHomeSummary';
 import { useFriendPublicLetters, useOtherPublicLetters } from '@/hooks/usePublicLetters';
